Restore working order for access control. Upgrade dependencies
