+++ packages/neoBackend/package.json
@@ -13,9 +13,9 @@
   "author": "Datavis Tech INC",
   "license": "UNLICENSED",
   "devDependencies": {
-    "eslint": "^6.5.1",
+    "eslint": "^6.8.0",
     "esm": "^3.2.25",
-    "prettier": "^1.18.2"
+    "prettier": "^1.19.1"
   },
   "dependencies": {
     "@teamwork/websocket-json-stream": "^2.0.0",
@@ -25,10 +25,10 @@
     "cookie-parser": "^1.4.4",
     "express": "^4.17.1",
     "vizhub-controllers": "1",
-    "vizhub-entities": "1",
     "vizhub-database": "1",
+    "vizhub-entities": "1",
     "vizhub-server-gateways": "1",
     "vizhub-use-cases": "1",
-    "ws": "^7.2.0"
+    "ws": "^7.2.1"
   }
 }
