+++ packages/neoBackend/src/serveShareDB/shareDBMiddleware/identifyAgent.js
@@ -1,6 +1,10 @@
 import { parse } from 'cookie';
 import { getUserIDFromJWT } from 'vizhub-controllers';
 
+// Populates request.agent.userId or request.agent.isServer.
+//
+// This ShareDB middleware triggers when new connections are made,
+// whether from the browser or from the server.
 export const identifyAgent = (request, done) => {
   // If the connection is coming from the browser,
   if (request.req) {
