+++ packages/neoFrontend/package.json
@@ -15,22 +15,22 @@
   "license": "UNLICENSED",
   "dependencies": {
     "@datavis-tech/ot-json0": "^2.1.0",
-    "@teamwork/sharedb": "^3.1.0",
-    "codemirror": "^5.49.2",
+    "codemirror": "^5.50.0",
     "d3-color": "^1.4.0",
     "diff-match-patch": "^1.0.4",
-    "immer": "^4.0.2",
+    "immer": "^5.1.0",
     "is-mobile": "^2.1.0",
     "json0-ot-diff": "^1.0.5",
-    "marked": "^0.7.0",
+    "marked": "^0.8.0",
     "perfect-scrollbar": "^1.4.0",
-    "query-string": "^6.8.3",
-    "react": "16.11.0",
-    "react-dom": "16.11.0",
+    "query-string": "^6.9.0",
+    "react": "16.12.0",
+    "react-dom": "16.12.0",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.2.0",
-    "styled-components": "^4.4.0",
+    "react-scripts": "^3.3.0",
+    "sharedb": "^1.1.0",
+    "styled-components": "^4.4.1",
     "vizhub-entities": "1",
     "vizhub-presenters": "1"
   },
@@ -41,6 +41,6 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "prettier": "^1.18.2"
+    "prettier": "^1.19.1"
   }
 }
