+++ packages/neoFrontend/src/pages/VizPage/VizContext/useShareDBDoc.js
@@ -1,10 +1,16 @@
-import { useEffect, useState, useContext } from 'react';
+import { useEffect, useState, useContext, useCallback } from 'react';
 import { ConnectionContext } from '../ConnectionContext';
+import { WarningContext } from '../WarningContext';
 
 export const useShareDBDoc = (collection, id) => {
   const connection = useContext(ConnectionContext);
+  const { setWarning } = useContext(WarningContext);
 
   const [vizContentDoc, setVizContentDoc] = useState(null);
+
+  const onError = useCallback(error => {
+    setWarning(error.message);
+  }, [setWarning]);
 
   useEffect(() => {
     if (!connection) return;
@@ -13,7 +19,10 @@ export const useShareDBDoc = (collection, id) => {
     // (don't want the user to see stale stuff).
     setVizContentDoc(null);
 
-    const doc = connection.get(collection, id);
+    const doc = connection.get(collection, id, err => {
+      console.log('hered');
+    });
+    doc.on('error', onError);
 
     // Wait until subscribe finishes before passing this out,
     // so that downstream code can assume data is present
@@ -23,9 +32,10 @@ export const useShareDBDoc = (collection, id) => {
     });
 
     return () => {
+      doc.off('error', onError);
       doc.unsubscribe();
     };
-  }, [connection, collection, id]);
+  }, [connection, collection, id, onError]);
 
   return vizContentDoc;
 };
