+++ packages/serverGateways/package.json
@@ -3,17 +3,16 @@
   "version": "1.0.0",
   "description": "",
   "main": "src/index.js",
-  "scripts": {
-  },
+  "scripts": {},
   "author": "",
   "license": "UNLICENSED",
   "dependencies": {
     "@datavis-tech/ot-json0": "^2.1.0",
-    "@teamwork/sharedb": "^3.1.0",
-    "@teamwork/sharedb-mingo-memory": "^3.0.1",
-    "@teamwork/sharedb-mongo": "^3.0.0",
     "mongodb": "^3.3.3",
-    "vizhub-gateways": "1",
-    "vizhub-database": "1"
+    "sharedb": "^1.1.0",
+    "sharedb-mingo-memory": "^1.1.1",
+    "sharedb-mongo": "^1.0.0-beta.11",
+    "vizhub-database": "1",
+    "vizhub-gateways": "1"
   }
 }
